Cancel outstanding channel requests if connection to the CM is lost.
When the connection to a given channel manager was lost (such as it crashed),
any outstanding channel requests were freed, but then the request callback
fired with an invalid (already freed) DispatcherRequestData parameter. It
subsequently used this invalid data and crashed.

This patch cancels all outstanding channel requests when they are freed so the
callback isn't called with invalid data. Fixes #598332