account_assistant_apply_account_and_finish: explicitely ask to set the display name
Checking if priv->current_widget_object is NULL or not isn't right because if
we create a normal account before the Salut one, this pointer is not NULL and
we rename the Salut account to the same name that the standard one.

This patch adds an extra argument account_assistant_apply_account_and_finish
saying if we should set a new display name or not. It's more clear and avoid
to rely on side effects.

Fixes bug #615058