src/empathy-chat-window.ui: Swap keyboard controls for moving tabs
Closes: bgo#615164
Keyboard shortcuts for moving tabs and for changing tabs work in the opposite direction.
Users may expect them to work in the same direction and to be consistant with other GNOME applications.
The mapping is now Ctrl + Shift + PageUp / PageDown for moving tabs left / right respectively, as it is in GNOME Terminal for instance.