Main window: mark Edit→Context as untranslatable.
The label of this menu item is changed to either _("Contact") or
_("Group"), depending on what's currently selected in the tree view. If
neither a contact nor a group is selected, the item is hidden entirely.

Thus, it should be safe not to translate this string.

I haven't tested this. I just came across this when mpt pointed it out
on the Twitter: <http://twitter.theinfo.org/30701150579720192>. His
tweets suggest that he's found a way to make it show up in the UI, but I
can't find any way to do that (either reading the code, or playing with
the actual UI).

https://bugzilla.gnome.org/show_bug.cgi?id=640751