Add a Remember Password toggle which remembers the password
This feels a bit awkwardly implemented, for example, if you enter a password
but uncheck the toggle, then the password will be removed from the dialog when
the account connects (maybe this is the correct behaviour?). No one has really
specced out the flow of this, maybe it can be sorted out properly when this
code is made more generic.