Really remove user from the blocked list when readding them
Previous commit in 161c9a2381a9c2df45ef8df81e8b680ce41b5c64 would have worked,
if not for Folks. And it turns out the CM I tested against already happens to
forget the blocked status if you readd a contact, so it looked like it was
working.

Fixes: https://bugzilla.gnome.org/show_bug.cgi?id=645487