Bug 646227 — Possible overflow in persona-view:drag_data_get
I think it was a fluke that this code worked in the first place, since it was
accessing arrays with indices which were out of range, and which wouldn't have
corresponded to the correct entries even if they were in range.

This should fix the mapping between DND types and DND URIs/atoms in the
contact list, individual and persona views.

Closes: bgo#646227