Include SMSChannel property in empathy_chat_window_find_chat()
The approach in https://bugzilla.gnome.org/show_bug.cgi?id=644085 was wrong
because channel paths can change across reconnects. This would cause new chat
windows to be created when they were rerequested.

A more future-proof solution would be to keep a white/black list of comparable
properties, but I think this is overengineering the problem for the time
being. Instead we'll just include the SMSChannel property in the comparison.