Use pixbuf for PixbufCellRenderer in protocol chooser
If an icon-name is given to a PixbufCellRenderer it will re-load the
image from the icon theme every time the size gets requested for a
different cell. Which happens a *LOT*.

Instead load the icon the row is added to the model, such that the
cell renderer can simply request the size of the pixbuf. This massive
speeds up the performance of the protocol chooser combobox.