roster-view: don't display offline not favorite top contacts
We just want to always display the favorites offline contacts in the top
contact section.

One may say that we are kinda breaking the model group abstraction, which is
probably true. Ideally we may want to have a roster view subclass implementing
the top contact logic but life is too short.

https://bugzilla.gnome.org/show_bug.cgi?id=683022