roster-view: make sure the same contact isn't added twice to the same group
Calling empathy_roster_model_get_groups_for_individual() in individual_added
may have the side effect for folks to discover the contact's groups and so
fire a bunch of signals leading us to add the contact in the group.
Then we look for his groups in individual_added() and re-add him a second
time.

Fix this by making sure we won't add the same contact twice to the same group.

https://bugzilla.gnome.org/show_bug.cgi?id=683025