Make parallel builds work
There is no need to execute glib-ginterface-gen.py twice because it
generates both the .c and .h files in one shot. Therefore, merely
having one of them (say the .c) depend on the other (say the .h) is
enough.

Moreover, running it twice breaks parallel builds because one
glib-ginterface-gen.py process steps on the other.

Based on a suggestion from Ryan Lortie.

Fixes: https://bugzilla.gnome.org/688132