remove 'Search for contacts' accelerator
It was impossible to type 'F' in our status message because of it.

Shift+F is not a very common accelerator and there is no way to discover it
any more (accelerators are not displayed in Shell's menu).
Searching for contacts isn't something you are supposed to do every day so not
having an accelerator isn't that bad anyway.

https://bugzilla.gnome.org/show_bug.cgi?id=687629