account-settings: early return if SASL earlier
We should update the URI schemes and service even if the account uses SASL, so
it has to be done *before* the early return.

I did not hit this bug before as the only CM on which we use URI schemes
(rakia) was not implementing SASL channel when I implemented this feature.

https://bugzilla.gnome.org/show_bug.cgi?id=691195