account-settings: don't set the presence on newly created accounts
What happened before this patch was:
1) tpaw_account_settings_do_create_account() sets the requested presence
   to the global requested presence in Empathy;
2) If the requested presence is unset, unknown or offline,
   empathy_connect_new_account() sets it to available;
3) Every process with a TpAccountManager object in memory that was
   previously used to set the requested presence will try to set the
   presence on the new account.

The first step is a problem as it doesn't work if the account is created
by a process that is not Empathy as the other processes have no way of
knowing what Empathy's requested presence is.
This doesn't matter as, in the third step, both Empathy and gnome-shell
(and any other process using a TpAccountManager) will try to set the
presence to their own idea of the current requested presence. Of course
this is a race condition, see [1], but the behaviour is not changed by
this patch.
In case there is no process using a TpAccountManager, the second step
will still guarantee that the newly created account goes online.

[1] https://bugs.freedesktop.org/show_bug.cgi?id=24104

https://bugzilla.gnome.org/show_bug.cgi?id=699492